manifests: split waitable object
waitable objects are needed by the commandline
flow, not by packages proper.

Until a better API emerges, isolate the related
code in a separate package.
